refactor: AddItemView와 EditItemView를 ItemView로 통합, receiveData() 메서드 ItemView로 위치 수정, view의 UI 프로퍼티 접근 제어자 수정
